Stop board processing if a winner is found
Added a check to halt board processing loops when a winner is identified. This fixes a bug introduced when board processing was changed to no longer short-circuit when a winner is declared (which was in support of animating the chain reactions). Updated game logic, main script, and test cases to incorporate this change.